Move some file to be logic
