+++ scripts/install/install_docopt.sh
@@ -16,6 +16,6 @@ fi
 cd ${QPACKAGE_ROOT}
 
 rm -f -- scripts/${DOCOPT}{,c}
-${QPACKAGE_ROOT}/scripts/fetch_from_web.py ${DOCOPT_URL} ${DOCOPT}
+${QPACKAGE_ROOT}/scripts/install/fetch_from_web.py ${DOCOPT_URL} ${DOCOPT}
 
-mv ${DOCOPT} scripts/${DOCOPT}
+mv ${DOCOPT} scripts/utility/${DOCOPT}
