+++ scripts/install/install_emsl.sh
@@ -15,7 +15,7 @@ fi
 
 cd ${QPACKAGE_ROOT}
 
-${QPACKAGE_ROOT}/scripts/fetch_from_web.py ${URL} ${QPACKAGE_ROOT}/${BASE}.tar.gz
+${QPACKAGE_ROOT}/scripts/install/fetch_from_web.py ${URL} ${QPACKAGE_ROOT}/${BASE}.tar.gz
 tar -zxf ${BASE}.tar.gz && rm ${BASE}.tar.gz ||exit 1
 rm -rf EMSL_Basis
 mv ${BASE}-master EMSL_Basis
