+++ scripts/install/install_ezfio.sh
@@ -16,7 +16,7 @@ fi
 cd ${QPACKAGE_ROOT}
 
 rm -rf -- EZFIO
-${QPACKAGE_ROOT}/scripts/fetch_from_web.py ${URL} ${QPACKAGE_ROOT}/${BASE}.tar.gz
+${QPACKAGE_ROOT}/scripts/install/fetch_from_web.py ${URL} ${QPACKAGE_ROOT}/${BASE}.tar.gz
 tar -zxf ${BASE}.tar.gz && rm ${BASE}.tar.gz ||exit 1
 mv EZFIO-master EZFIO
