+++ scripts/install/install_irpf90.sh
@@ -17,7 +17,7 @@ fi
 cd ${QPACKAGE_ROOT}
 
 
-${QPACKAGE_ROOT}/scripts/fetch_from_web.py ${URL} ${QPACKAGE_ROOT}/${BASE}.tar.gz
+${QPACKAGE_ROOT}/scripts/install/fetch_from_web.py ${URL} ${QPACKAGE_ROOT}/${BASE}.tar.gz
 tar -zxf ${BASE}.tar.gz && rm ${BASE}.tar.gz ||exit 1
 mv ${BASE}-master irpf90
 make -C irpf90 | tee install_irpf90.log
