+++ scripts/install/install_m4.sh
@@ -19,7 +19,7 @@ cd ${QPACKAGE_ROOT}
 if [[ -z ${M4} ]] 
 then 
     rm -f -- bin/m4
-    ${QPACKAGE_ROOT}/scripts/fetch_from_web.py ${M4_URL} M4.tar.gz 
+    ${QPACKAGE_ROOT}/scripts/install/fetch_from_web.py ${M4_URL} M4.tar.gz 
     tar -zxf M4.tar.gz && rm M4.tar.gz ||exit 1
     cd m4* || exit 1 
     ./configure && make || exit 1
