+++ scripts/install/install_resultsFile.sh
@@ -17,7 +17,7 @@ cd ${QPACKAGE_ROOT}
 
 
 rm -rf resultsFile-master
-${QPACKAGE_ROOT}/scripts/fetch_from_web.py ${URL} ${QPACKAGE_ROOT}/resultsFile.tar.gz
+${QPACKAGE_ROOT}/scripts/install/fetch_from_web.py ${URL} ${QPACKAGE_ROOT}/resultsFile.tar.gz
 tar -zxf resultsFile.tar.gz && rm resultsFile.tar.gz ||exit 1
 mv resultsFile-master resultsFile
